Optimalize, improve readibility. Rename PacketCreator to packet_creator

+++ minecraft_console_client/data_structures/position.py
@@ -4,7 +4,7 @@ from typing import Union
 class Position:
     pos = {'x': None, 'y': None, 'z': None}
 
-    def __init__(self, pos: (int, int, int)):
+    def __init__(self, pos: (float, float, float)):
         """ pos = (x, y, z)"""
         self.pos['x'] = pos[0]
         self.pos['y'] = pos[1]
@@ -15,5 +15,4 @@ class Position:
         return pos['x'], pos['y'], pos['z']
 
     def __repr__(self):
-
         return f"[x: {self.pos['x']}, y: {self.pos['y']}, z: {self.pos['z']}]"
